+++ _pylibczi/Reader.cpp
@@ -2,7 +2,7 @@
 // Created by Jamie Sherman on 2019-08-18.
 //
 #include <tuple>
-#include "aics_added.h"
+#include "Reader.h"
 #include "exceptions.h"
 
 namespace pylibczi {
